like-function add

+++ app/controllers/likes_controller.rb
@@ -1,18 +1,44 @@
 class LikesController < ApplicationController
   
   def create
-    @like = current_company.likes.build(like_params)
+   
     @product = Product.find_by(id: params[:product_id])
-    if @like.save
-     redirect_to @product
+    ### current_collegeとcurrent_comapnyの分岐
+    if current_college
+      @like = current_college.likes.build(like_params)
+      @like.company_id = nil 
+      if @like.save!
+         redirect_to @product
+      else
+         redirect_to @product
+      end 
+    elsif 
+      @like = current_company.likes.build(like_params)
+       @like.college_id = nil 
+      if @like.save!
+         redirect_to @product
+      else
+        redirect_to @product
+      end 
     end
   end
+
 
   def destroy
     @like = Like.find_by(id: params[:id])
     @product = Product.find_by(id: params[:product_id])
-    if @like.destroy
-       redirect_to @product
+    if current_college
+        if @like.destroy
+         redirect_to @product
+        else
+        redirect_to @product
+        end
+    else 
+        if @like.destroy
+         redirect_to @product
+        else
+        redirect_to @product
+        end
     end
   end
   
@@ -22,6 +48,5 @@ class LikesController < ApplicationController
    def like_params
      params.permit(:product_id)
    end
-   
-end
+ end
